Moved DUP departures logic to new submodule.
